+++ src/main/kotlin/dev/nyon/skylper/skyblock/menu/Menu.kt
@@ -1,0 +1,9 @@
+package dev.nyon.skylper.skyblock.menu
+
+import dev.nyon.skylper.skyblock.menu.collections.Collections
+
+object Menu {
+    fun init() {
+        Collections.init()
+    }
+}